pyproject.toml
[project]
name = "homesignal-ai"
version = "0.1.0"
description = "동대문구 부동산 시계열 예측 및 RAG 챗봇 서비스"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.109.0",
    "uvicorn[standard]>=0.27.0",
    "pydantic>=2.0",
    "pydantic-settings>=2.0",
    "supabase>=2.0",
    "pyiceberg>=0.5.0",  # supabase storage3 analytics 의존성
    "redis>=5.0",
    "httpx>=0.26.0",
    "openai>=1.10.0",
    "anthropic>=0.18.0",
    "pyyaml>=6.0",
    "psycopg[binary]>=3.3.3",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "ruff>=0.2",
    "mypy>=1.8",
]
ml = [
    "prophet>=1.1",
    "lightgbm>=4.0",
    "pandas>=2.0",
    "numpy>=1.26",
    "scikit-learn>=1.4",
]
test = [
    "pytest>=8.0",
    "pytest-asyncio>=0.23",
    "pytest-cov>=4.0",
    "pytest-mock>=3.12",
]
crawler = [
    "beautifulsoup4>=4.12.0",
    "lxml>=5.0.0",
]

[tool.ruff]
target-version = "py311"
line-length = 88

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]

[tool.mypy]
python_version = "3.11"
strict = true
